rest_api/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
